% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{calculate_daily_temp_unit}
\alias{calculate_daily_temp_unit}
\title{Calculate Daily Temperature Unit}
\usage{
calculate_daily_temp_unit(
  average_temp,
  temp_min,
  temp_max,
  ideal_temp_min,
  ideal_temp_max
)
}
\arguments{
\item{average_temp}{Average daily temperature or vector of temperatures.}

\item{temp_min}{Minimum viable growing temperature.}

\item{temp_max}{Maximum viable growing temperature.}

\item{ideal_temp_min}{Lower bound of ideal growing window.}

\item{ideal_temp_max}{Upper bound of ideal growing window.}
}
\value{
Daily temperature unit.
}
\description{
The daily temperature unit is used to modify growth speed and efficiency.
All temperature are in Celsius.
}
