% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{german_soil}
\alias{german_soil}
\title{Compute iCrop2 compatible soil values from german soil info}
\usage{
german_soil(
  name,
  GPV,
  LK,
  nFK,
  kf,
  we,
  depth_top_layer = 200,
  albedo = 0.1,
  drain_factor = NULL,
  sat_drain_factor = 0.5,
  slope = 10,
  hydration_depth = 600
)
}
\arguments{
\item{we}{Effektive Wurzelungstiefe in cm.}

\item{drain_factor}{Percentage of water over ATSW that drains per day. If not set
a crude linear model will be used. Generally related to soil composition:
more clay = smaller df (~0.2 for predominantly clay), more sand = larger df (~0.75
for predom. sand). See Sinclair & Soltani (2012) p. 177, table 14.2}
}
\description{
Units mm/mm so divide by 100 for mm/dm
}
