% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCrop2R-package.R
\docType{package}
\name{iCrop2R-package}
\alias{iCrop2R}
\alias{iCrop2R-package}
\title{iCrop2R: Implementation Of The SSM-iCrop2 Crop Modell}
\description{
An R implementation of the SSM-iCrop2 crop model.
}
\author{
\strong{Maintainer}: Jacob Wujciak \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID})

}
\keyword{internal}
