% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{initialize_weather}
\alias{initialize_weather}
\title{Update Weather}
\usage{
initialize_weather(sim_env, temp_mod = 0, rain_mod = 1)
}
\arguments{
\item{sim_env}{Simulation environment created with \code{\link[=create_sim_env]{create_sim_env()}}.}

\item{temp_mod}{Additive modifier for temperature.}

\item{rain_mod}{Multiplicative modifier for precipitation.}
}
\value{
Updated weather data.
}
\description{
Updates the weather with modifiers and calculates snow and snow melt.
}
