% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{kf2CN}
\alias{kf2CN}
\title{convert gesättigte Wasserleitfähigkeit to CN runoff curve number}
\usage{
kf2CN(kf)
}
\arguments{
\item{kf}{gesättigte Wasserleitfähigkeit in cm/d}
}
\description{
convert gesättigte Wasserleitfähigkeit to CN runoff curve number
}
