% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{prepare_weather}
\alias{prepare_weather}
\title{Update Weather}
\usage{
prepare_weather(weather, crop, albedo, rain_mod = 1, temp_mod = 0)
}
\arguments{
\item{rain_mod}{Multiplicative modifier for precipitation.}

\item{temp_mod}{Additive modifier for temperature.}

\item{sim_env}{Simulation environment created with \code{\link[=create_sim_env]{create_sim_env()}}.}
}
\value{
Updated weather data.
}
\description{
Updates the weather with modifiers and calculates snow and snow melt.
}
