% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Simulation}
\alias{Simulation}
\title{Simulate crop growth}
\description{
This class implements the iCrop2 model.
}
\seealso{
https://doi.org/10.1016/j.agsy.2020.102855
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simulation-new}{\code{Simulation$new()}}
\item \href{#method-Simulation-prepare_weather}{\code{Simulation$prepare_weather()}}
\item \href{#method-Simulation-get_data}{\code{Simulation$get_data()}}
\item \href{#method-Simulation-get_state}{\code{Simulation$get_state()}}
\item \href{#method-Simulation-get_fallow_data}{\code{Simulation$get_fallow_data()}}
\item \href{#method-Simulation-get_sowing_date}{\code{Simulation$get_sowing_date()}}
\item \href{#method-Simulation-run_simulation}{\code{Simulation$run_simulation()}}
\item \href{#method-Simulation-reset}{\code{Simulation$reset()}}
\item \href{#method-Simulation-clone}{\code{Simulation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$new(
  crop,
  location,
  management,
  soil,
  weather,
  hydration_depth = 600
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{crop}}{A list containing the variable-values pairs used in the 'crop'-sheet of the excel version.
This package comes with a version of the iCrop2 crop database \link{crops}.}

\item{\code{location}}{A list containing the variable-values pairs used in the 'location'-sheet of the excel version.
(Only the actual variables, not sheet names or row numbers.)}

\item{\code{management}}{A list containing the variable-values pairs used in the 'Management'-sheet of the excel version.}

\item{\code{soil}}{A list containing the variable-values pairs used in the 'soil'-sheet of the excel version.
Can be generated from german-style soil profiles with \code{\link[=german_soil]{german_soil()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-prepare_weather"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-prepare_weather}{}}}
\subsection{Method \code{prepare_weather()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$prepare_weather()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-get_state"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-get_state}{}}}
\subsection{Method \code{get_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$get_state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-get_fallow_data"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-get_fallow_data}{}}}
\subsection{Method \code{get_fallow_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$get_fallow_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-get_sowing_date"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-get_sowing_date}{}}}
\subsection{Method \code{get_sowing_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$get_sowing_date()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-run_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-run_simulation}{}}}
\subsection{Method \code{run_simulation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$run_simulation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$reset(sure = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Simulation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
