% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{srad_from_sunh}
\alias{srad_from_sunh}
\title{Calculate solar radiation from sunshine hours}
\usage{
srad_from_sunh()
}
\arguments{
\item{lat}{Latitude in degrees. South latitudes should be entered
as negative.}
}
\description{
Calculate solar radiation from sunshine hours
}
