% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCrop2R-package.R
\docType{package}
\name{iCrop2R-package}
\alias{iCrop2R}
\alias{iCrop2R-package}
\title{iCrop2R: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
