% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{prepare_weather}
\alias{prepare_weather}
\title{Update Weather}
\usage{
prepare_weather(weather, crop, albedo, rain_mod = 1, temp_mod = 0)
}
\arguments{
\item{weather}{Weather data.}

\item{crop}{Crop data.}

\item{rain_mod}{Multiplicative modifier for precipitation.}

\item{temp_mod}{Additive modifier for temperature.}
}
\value{
Updated weather data.
}
\description{
Updates the weather with modifiers and calculates snow and snow melt.
}
